requirements.txt
praw
pandas
newspaper3k
sumy